+++ tv.sh
@@ -2,15 +2,15 @@
 
 LOGTO=/media/35a8/home/londonfire93/private/deluge/Scripts/logs/log.txt
 LOCKFILE=/media/35a8/home/londonfire93/private/deluge/lock.tv
-CURRENT_PID=$(cat /media/35a8/home/londonfire93/.config/deluge/deluged.pid|sha1sum|cut -c 1-10)
+CURRENT_PID=$(< /media/35a8/home/londonfire93/.config/deluge/deluged.pid|sha1sum|cut -c 1-10)
 
-if [ -e ${LOCKFILE} ] && kill -0 `cat ${LOCKFILE}`; then
+if [ -e ${LOCKFILE} ] && kill -0 "cat ${LOCKFILE}"; then
     echo "already running"
     exit
 fi
 
 # make sure the lockfile is removed when we exit and then claim it
-trap "rm -f ${LOCKFILE}; exit" INT TERM EXIT
+trap 'rm -f ${LOCKFILE}; exit' INT TERM EXIT
 echo $$ > ${LOCKFILE}
 
 # Process Files 
@@ -23,7 +23,7 @@ echo "$TVPATH $TVFOLDER $OUTPATH $CREATION_PID"
 		find "$TVPATH/$TVFOLDER" -name "*.rar" -exec 7z x "{}" -y -o"$TVPATH" \;
 		mkdir -p  "${OUTPATH}"
 		find "$TVPATH" -type f \( -iname \*.mkv -o -iname \*.mp4 -o -iname \*.avi \) -exec mv "{}" "${OUTPATH}" \;
-		rm -rf "$TVPATH/$TVFOLDER"
+		rm -rf "${TVPATH:?}/${TVFOLDER:?}"
 		echo "Done  $(date) $TVPATH/$TVFOLDER" >>  ${LOGTO}
 	fi
     fi
